git chore(es6): Refactor to ES6 modules for UMD bundle
