+++ .gitignore
@@ -1,7 +1,8 @@
 dist
 node_modules
-examples/querybase.js
-examples/querybase.min.js
+examples/**/*.js
 examples/**/*.d.ts
 coverage
-tsd.json
+es6
+.DS_Store
+.vscode/
