+++ .npmignore
@@ -2,3 +2,4 @@ node_modules
 logos
 examples
 coverage
+es6
