+++ examples/index.html
@@ -3,7 +3,7 @@
 
 <head>
 	<script src="https://www.gstatic.com/firebasejs/live/3.0/firebase.js"></script>
-	<script src="querybase.js"></script>
+	<script src="querybase.umd.js"></script>
 	<meta charset="utf-8">
 	<title>Querybase</title>
 </head>
