+++ tests/firebaseServer.js
@@ -1,5 +1,3 @@
-/// <reference path="../typings/tsd.d.ts" />
-
 const firebase = require("firebase");
 
 /**
