+++ tests/helpers.d.ts
@@ -1,10 +1,9 @@
-/// <reference path="../typings/tsd.d.ts" />
 
-interface Helpers {
-  isFunction(fn: () => {}): Boolean;
-  isFirebaseRef(ref: Firebase): Boolean;
-}
+// interface Helpers {
+//   isFunction(fn: () => {}): Boolean;
+//   isFirebaseRef(ref: Firebase): Boolean;
+// }
 
-declare module 'TestHelpers' {
-	export = Helpers;
-}
+// declare module 'TestHelpers' {
+// 	export = Helpers;
+// }
