--- typings/firebaseServer.d.ts
@@ -1,15 +1,0 @@
-declare module 'firebaseServer' {
-  
-  interface firebaseServer {
-    (config: FirebaseServiceConfig);
-  }
-
-  interface FirebaseServiceConfig {
-    url: String;
-    port: Number;
-    data: Object;
-  }
-  
-  export = firebaseServer;
-  
-}
