--- typings/tsd.d.ts
@@ -1,2 +1,0 @@
-/// <reference path="firebase/firebase.d.ts" />
-/// <reference path="node/node.d.ts" />
